matched Actor00100_Fn01900 3
Ports the beam-draw body from its matched siblings func_actor_405800_801329C8
and func_actor_400600_80132294; the parent coordinate is reached from the
Gfx_ViewWorldMtx symbol with OFFSET_OF so the la matches the ROM's relocation,
and the prototype's first three arguments are s16 (the callee narrows them).

+++ include/actors/actor_400100_update.h
@@ -36,7 +36,7 @@ extern Actor00100AnimWord    Actor00100_D1B9BC, Actor00100_D1B9C0;
 extern s8                    D_8007218A;
 extern u8                    D_80073BA9, D_801153F4;
 extern GsCOORDINATE2         Gfx_ViewCoord;
-void                         Actor00100_Fn01900(Actor00100*, s32, s32, s32, s32, s32);
+void                         Actor00100_Fn01900(Actor00100*, s16, s16, s16, s16, u8);
 void                         Actor00100_Fn0375C(Actor00100*);
 
 static __inline__ s16 Actor00100_InRegion(Actor00100* actor)
